+++ tests/unit/classes/classes/FooClass.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace mako\tests\unit\classes\classes;
+
+abstract class FooClass implements FooInterface
+{
+
+}