+++ tests/unit/classes/classes/FooInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace mako\tests\unit\classes\classes;
+
+interface FooInterface
+{
+
+}