+++ tests/unit/classes/classes/FooTrait.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace mako\tests\unit\classes\classes;
+
+trait FooTrait
+{
+
+}