+++ tests/unit/classes/preload/classes/CB.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace mako\tests\unit\classes\preload\classes;
+
+class CB extends CA
+{
+
+}