+++ tests/unit/classes/preload/classes/CD.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace mako\tests\unit\classes\preload\classes;
+
+class CD
+{
+	use TA;
+}