+++ tests/unit/preload/classes/CC.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace mako\tests\unit\preload\classes;
-
-class CC implements IA
-{
-
-}