+++ tests/unit/preload/classes/CD.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace mako\tests\unit\preload\classes;
-
-class CD
-{
-	use TA;
-}