+++ tests/unit/preload/classes/TA.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace mako\tests\unit\preload\classes;
-
-trait TA
-{
-
-}